Move initRepo to test
