+++ porch/repository/pkg/git/testing.go
@@ -492,107 +492,6 @@ func (p *objectWalker) walkObjectTree(hash plumbing.Hash) error {
 	return nil
 }
 
-// initRepo is a helper that creates a first commit, ensuring the repo is not empty.
-func initRepo(repo *gogit.Repository) error {
-	store := repo.Storer
-
-	var objectHash plumbing.Hash
-	{
-		data := []byte("This is a test repo")
-		eo := store.NewEncodedObject()
-		eo.SetType(plumbing.BlobObject)
-		eo.SetSize(int64(len(data)))
-
-		w, err := eo.Writer()
-		if err != nil {
-			return fmt.Errorf("error creating object writer: %w", err)
-		}
-
-		if _, err = w.Write(data); err != nil {
-			w.Close()
-			return fmt.Errorf("error writing object data: %w", err)
-		}
-		if err := w.Close(); err != nil {
-			return fmt.Errorf("error closing object data: %w", err)
-		}
-
-		if h, err := store.SetEncodedObject(eo); err != nil {
-			return fmt.Errorf("error storing object: %w", err)
-		} else {
-			objectHash = h
-		}
-	}
-
-	var treeHash plumbing.Hash
-	{
-		tree := object.Tree{}
-
-		te := object.TreeEntry{
-			Name: "README.md",
-			Mode: filemode.Regular,
-			Hash: objectHash,
-		}
-		tree.Entries = append(tree.Entries, te)
-
-		eo := store.NewEncodedObject()
-		if err := tree.Encode(eo); err != nil {
-			return fmt.Errorf("error encoding tree: %w", err)
-		}
-		if h, err := store.SetEncodedObject(eo); err != nil {
-			return fmt.Errorf("error storing tree: %w", err)
-		} else {
-			treeHash = h
-		}
-	}
-
-	var commitHash plumbing.Hash
-	{
-		now := time.Now()
-		commit := &object.Commit{
-			Author: object.Signature{
-				Name:  "Porch Author",
-				Email: "author@kpt.dev",
-				When:  now,
-			},
-			Committer: object.Signature{
-				Name:  "Porch Committer",
-				Email: "committer@kpt.dev",
-				When:  now,
-			},
-			Message:  "First commit",
-			TreeHash: treeHash,
-		}
-
-		eo := store.NewEncodedObject()
-		if err := commit.Encode(eo); err != nil {
-			return fmt.Errorf("error encoding commit: %w", err)
-		}
-		if h, err := store.SetEncodedObject(eo); err != nil {
-			return fmt.Errorf("error storing commit: %w", err)
-		} else {
-			commitHash = h
-		}
-	}
-
-	{
-		ref := plumbing.NewHashReference(Main, commitHash)
-		if err := repo.Storer.SetReference(ref); err != nil {
-			return fmt.Errorf("error setting reference %q: %w", Main, err)
-		}
-
-		// gogit uses suboptimal default reference name; delete it
-		repo.Storer.RemoveReference(plumbing.Master)
-
-		// create correct HEAD as a symbolic reference of main branch
-		head := plumbing.NewSymbolicReference(plumbing.HEAD, Main)
-		if err := repo.Storer.SetReference(head); err != nil {
-			return fmt.Errorf("error creating HEAD ref: %w", err)
-		}
-	}
-
-	return nil
-}
-
 // parseHash is a helper that parses a GitHash provided by the client.
 func parseHash(s string) (GitHash, error) {
 	var h GitHash
